Seperate dependency with composition components && Mod Logic
