+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export * from './Composition';
+export * from './CompositionTable';
